backend/pyproject.toml
[project]
name = "rag-overlawdocuemnts"
version = "0.1.0"
description = "The application backend"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.1.10",
    "langchain-huggingface>=0.3.1",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=5.1.0",
    "lancedb>=0.6.15",
    "docling>=2.51.0",
    "langchain-docling>=1.0.0",
    "loguru>=0.7.3",
    "fastapi",
    "pydantic",
    "uvicorn",
    "transformers @ git+https://github.com/huggingface/transformers.git@v4.56.0-Embedding-Gemma-preview",
]
